move props off of state.. oops...
